refactor: 기물에서 공통 이동 규칙을 추상 클래스로 분리
