+++ src/main/java/janggi/model/gimul/palace/Sa.java
@@ -1,0 +1,14 @@
+package janggi.model.gimul.palace;
+
+import janggi.model.Team;
+
+public class Sa extends AbstractPalaceGimul {
+    public Sa(Team team) {
+        super(team);
+    }
+
+    @Override
+    public String getSymbol() {
+        return "사";
+    }
+}
