+++ src/main/java/janggi/model/initializer/BoardInitializerWithConst.java
@@ -2,11 +2,11 @@ package janggi.model.initializer;
 
 import janggi.model.Team;
 import janggi.model.gimul.Byeong;
-import janggi.model.gimul.Cha;
+import janggi.model.gimul.linearMove.Cha;
 import janggi.model.gimul.Gimul;
-import janggi.model.gimul.Jang;
-import janggi.model.gimul.Pho;
-import janggi.model.gimul.Sa;
+import janggi.model.gimul.palace.Jang;
+import janggi.model.gimul.linearMove.Pho;
+import janggi.model.gimul.palace.Sa;
 import janggi.model.position.Column;
 import janggi.model.position.Position;
 import janggi.model.position.Row;
