+++ src/test/java/janggi/gimul/JangTest.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import janggi.model.Team;
-import janggi.model.gimul.Cha;
+import janggi.model.gimul.linearMove.Cha;
 import janggi.model.gimul.Gimul;
-import janggi.model.gimul.Jang;
+import janggi.model.gimul.palace.Jang;
 import janggi.model.position.Column;
 import janggi.model.position.Position;
 import janggi.model.position.PositionPath;
