+++ src/test/java/janggi/gimul/SangTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import janggi.model.Team;
-import janggi.model.gimul.Cha;
+import janggi.model.gimul.linearMove.Cha;
 import janggi.model.gimul.Gimul;
 import janggi.model.gimul.Sang;
 import janggi.model.position.Column;
